pyproject.toml
[project]
name = "course-paper"
version = "0.1.0"
description = ""
authors = [
    {name = "IDaniil228",email = "99666220+IDaniil228@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (==4.2)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "requests (>=2.33.1,<3.0.0)",
    "pandas (>=3.0.2,<4.0.0)",
    "numpy (>=2.4.4,<3.0.0)",
    "django-filter (<25)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
